requirements.txt
flask
click
requests
python-dotenv
PyJWT